refactor : #129 GlobalProfile 컴포넌트 분리 및 WriterProfilePicDiv 삭제
